requirements.txt
anthropic==0.91.0
chromadb==1.5.7
faker==40.13.0
fastapi==0.135.3
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.6
pypdf==6.9.2
python-dotenv==1.2.2
sentence-transformers==5.3.0
streamlit==1.56.0
uvicorn==0.44.0
